+++ Shared/SharedUI/Shared.UiTest/BaseDialogs/PackFileTree/ContextMenu/Commands/SavePackFileContainerCommandTests.cs
@@ -6,7 +6,7 @@ using Shared.Core.Settings;
 using Shared.Ui.BaseDialogs.PackFileTree;
 using Shared.Ui.BaseDialogs.PackFileTree.ContextMenu.Commands;
 
-namespace Shared.UiTest.PackFileTree.ContextMenu.Commands
+namespace Shared.UiTest.BaseDialogs.PackFileTree.ContextMenu.Commands
 {
     [TestFixture]
     internal class SavePackFileContainerCommandTests : ContextMenuCommandTestBase
